:sparkles: Add create user, Login user updated and more.

+++ base/views.py
@@ -1,7 +1,10 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from django.contrib import messages
 from .models import Room, Topic
 from .forms import RoomForm
@@ -11,10 +14,14 @@ from .forms import RoomForm
 #    {'id':2, 'name':"Design with me."},
 #    {'id':3, 'name':"Just testing rooms."},
 #]
+
 def loginPage(request):
+    page = 'login' 
+    if request.user.is_authenticated:
+        return redirect('home')
 
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -32,14 +39,28 @@ def loginPage(request):
         else:
             messages.error(request, 'Password are not correct')
 
-    context = {}
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
 
+def registerPage(request):
+    form = UserCreationForm()
 
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+           user = form.save(commit=False) 
+           user.username = user.username.lower()
+           user.save()
+           login(request, user)
+           return redirect('home')
+        else:
+            messages.error(request, 'An error occurred during registration')
+
+    return render(request, 'base/login_register.html', {'form': form})
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -58,6 +79,8 @@ def room(request, pk):
     context = {'room':room}
     return render(request, 'base/room.html',context)
 
+
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
     
@@ -70,9 +93,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse("You're not the owner of this room")
+
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -85,8 +113,13 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You're not the owner of this room")
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
